+++ src/test/java/com/seollem/server/restdocs/controller/book/GetCalender.java
@@ -15,6 +15,7 @@ import static org.springframework.restdocs.request.RequestDocumentation.requestP
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.google.gson.Gson;
+import com.seollem.server.book.Book.BookStatus;
 import com.seollem.server.book.BookController;
 import com.seollem.server.book.BookMapper;
 import com.seollem.server.book.BookService;
@@ -43,7 +44,7 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
 @ExtendWith({RestDocumentationExtension.class, SpringExtension.class})
 @WebMvcTest(BookController.class)
-public class Abandon extends WebMvcTestSetUpUtil {
+public class GetCalender extends WebMvcTestSetUpUtil {
 
   @MockBean
   private GetEmailFromHeaderTokenUtil getEmailFromHeaderTokenUtil;
@@ -64,7 +65,7 @@ public class Abandon extends WebMvcTestSetUpUtil {
   private Gson gson;
 
   @Test
-  public void abandonTest() throws Exception {
+  public void calenderTest() throws Exception {
     //given
     given(getEmailFromHeaderTokenUtil.getEmailFromHeaderToken(Mockito.anyMap())).willReturn(
         "starrypro@gmail.com");
@@ -72,39 +73,39 @@ public class Abandon extends WebMvcTestSetUpUtil {
     given(memberService.findVerifiedMemberByEmail(Mockito.anyString())).willReturn(
         StubDataUtil.MockMember.getMember());
 
-    given(bookService.findAbandonedBooks(Mockito.anyInt(), Mockito.anyInt(),
-        Mockito.any(Member.class))).willReturn(
+    given(bookService.findVerifiedBooksByMemberAndBookStatus(Mockito.anyInt(), Mockito.anyInt(),
+        Mockito.any(Member.class), Mockito.any(BookStatus.class), Mockito.anyString())).willReturn(
         StubDataUtil.MockBook.getBookPage());
 
-    given(bookMapper.BooksToAbandonResponse(Mockito.any())).willReturn(
-        StubDataUtil.MockBook.getAbandonResponse());
+    given(bookMapper.BooksToCalenderResponse(Mockito.any())).willReturn(
+        StubDataUtil.MockBook.getCalenderResponse());
 
     //when
     ResultActions resultActions = mockMvc.perform(
-        MockMvcRequestBuilders.get("/books/abandon").accept(MediaType.APPLICATION_JSON)
+        MockMvcRequestBuilders.get("/books/calender").accept(MediaType.APPLICATION_JSON)
             .contentType(MediaType.APPLICATION_JSON).header("Authorization",
                 "Bearer eyJhbGciOiJIUzUxMiJ9.eyJyb2xlcyI6WyJST0xFX1VTRVIiXSwiZW1haWwiOiJzdGFycnlwcm9AZ21haWwuY29tIiwic3ViIjoic3RhcnJ5cHJvQGdtYWlsLmNvbSIsImlhdCI6MTY3OTQ2MTg0NSwiZXhwIjoxNjc5NDY3ODQ1fQ.ri2B7x4vhgydiJYdXtTTZuD9EZzX-l4coGwiPWYjYDUeWLSunsBDkdslQdzZb9D72qIlArzzP7nCalROkVYOTA")
             .queryParam("page", "1").queryParam("size", "10"));
 
     //then
-    resultActions.andExpect(status().isOk()).andDo(document("Abandon",
+    resultActions.andExpect(status().isOk()).andDo(document("Calender",
         preprocessRequest(modifyUris().scheme(SCHEMA).host(URI).removePort(), prettyPrint()),
         preprocessResponse(prettyPrint()),
         requestHeaders(headerWithName("Authorization").description("Bearer JWT Access Token")),
         requestParameters(
-            List.of(parameterWithName("page").description("원하는 page 수"),
+            List.of(parameterWithName("page").description("원하는 page"),
                 parameterWithName("size").description("페이지 별 책 개수"))),
         responseFields(
             fieldWithPath("item[].bookId").type(JsonFieldType.NUMBER).description("Book-id"),
-            fieldWithPath("item[].createdAt").type(JsonFieldType.STRING).description("책 등록 일자"),
-            fieldWithPath("item[].title").type(JsonFieldType.STRING).description("책 제목"),
             fieldWithPath("item[].cover").type(JsonFieldType.STRING).description("표지 이미지 url"),
+            fieldWithPath("item[].readEndDate").type(JsonFieldType.STRING).description("읽기 완료한 시간"),
             fieldWithPath("pageInfo.page").type(JsonFieldType.NUMBER).description("조회 페이지"),
             fieldWithPath("pageInfo.size").type(JsonFieldType.NUMBER).description("페이지별 책 개수"),
             fieldWithPath("pageInfo.totalElements").type(JsonFieldType.NUMBER)
                 .description("조회된 책의 총 개수"),
             fieldWithPath("pageInfo.totalPages").type(JsonFieldType.NUMBER)
                 .description("조회된 총 페이지 수")
+
         )));
   }
 }
